+++ build.sh
@@ -25,28 +25,17 @@ HELP="$0 [<target> ...] [<flag> ...]
    libclx - build the clx C++ code
    clx    - build the clx Python package
  and <flag> is:
-   -v         - verbose build mode
-   -g         - build for debug
-   -n         - no install step
-   --show_depr_warn - show cmake deprecation warnings
    -h         - print this text
 
- default action (no args) is to build and install 'libclx' then 'clx' targets
+ default action (no args) is to build and install 'clx' target
 "
-LIBCLX_BUILD_DIR=${REPODIR}/cpp/build
 CLX_BUILD_DIR=${REPODIR}/python/build
-BUILD_DIRS="${LIBCLX_BUILD_DIR} ${CLX_BUILD_DIR}"
-
-# Set defaults for vars modified by flags to this script
-VERBOSE=""
-INSTALL_TARGET=install
-BUILD_DISABLE_DEPRECATION_WARNING=ON
+BUILD_DIRS="${CLX_BUILD_DIR}"
 
 # Set defaults for vars that may not have been defined externally
 #  FIXME: if PREFIX is not set, check CONDA_PREFIX, but there is no fallback
 #  from there!
 INSTALL_PREFIX=${PREFIX:=${CONDA_PREFIX}}
-PARALLEL_LEVEL=${PARALLEL_LEVEL:=""}
 
 function hasArg {
     (( ${NUMARGS} != 0 )) && (echo " ${ARGS} " | grep -q " $1 ")
@@ -67,20 +56,6 @@ if (( ${NUMARGS} != 0 )); then
     done
 fi
 
-# Process flags
-if hasArg -v; then
-    VERBOSE=1
-fi
-if hasArg -g; then
-    BUILD_TYPE=Debug
-fi
-if hasArg -n; then
-    INSTALL_TARGET=""
-fi
-if hasArg --show_depr_warn; then
-    BUILD_DISABLE_DEPRECATION_WARNING=OFF
-fi
-
 # If clean given, run it prior to any other steps
 if hasArg clean; then
     # If the dirs to clean are mounted dirs in a container, the
@@ -96,23 +71,10 @@ if hasArg clean; then
 fi
 
 ################################################################################
-# Configure, build, and install libclx
-if (( ${NUMARGS} == 0 )) || hasArg libclx; then
-
-    mkdir -p ${LIBCLX_BUILD_DIR}
-    cd ${LIBCLX_BUILD_DIR}
-    cmake .. -DCMAKE_INSTALL_PREFIX=${INSTALL_PREFIX} -DDISABLE_DEPRECATION_WARNING=${BUILD_DISABLE_DEPRECATION_WARNING}
-    make -j${PARALLEL_LEVEL} VERBOSE=${VERBOSE} ${INSTALL_TARGET}
-fi
 
 # Build and install the clx Python package
 if (( ${NUMARGS} == 0 )) || hasArg clx; then
 
     cd ${REPODIR}/python
-    if [[ ${INSTALL_TARGET} != "" ]]; then
-        python setup.py build_ext --inplace
-        python setup.py install
-    else
-	    python setup.py build_ext --inplace --library-dir=${LIBCLX_BUILD_DIR}
-    fi
+    python setup.py install
 fi